requirements.txt
Django==5.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.3.0
whitenoise==6.8.2
djangorestframework==3.15.2

